[FAB-7831] Move funcs that propagate SDK context
This change moves SDK provides to a context object
that can be passed around. This removes the need to expose
all methods directly from the SDK object.

The underlying FabricClient method is also refactored
into a smaller Resource interface to limit exposure
from the SDK. This work continues in upcoming
changes.

Change-Id: I9233fd6278683cd2b24a47c6d4c10032e75f7c5f
Signed-off-by: Troy Ronda <troy@troyronda.com>