[FAB-7513] MSP-like key and cert storage
This change introduces a key value store that supports MSP-like
directory structure, the same that is currently used for
pre-enrolled users (generated by the cryptogen tool).
Also, the CredentialManager is modified to use this type of
store for user private keys (in case of SW crypto provider) and user certs.

Change-Id: I7e3dcd07d14354deb35514ee9275b88a5e9055ae
Signed-off-by: Aleksandar Likic <aleksandar.likic@securekey.com>