[FAB-8270] Split FabriCAClient interface
FabriCAClient interface is split into:
- EnrollmentService - users without signing identities
  use it to signup with a Fabric network
- IdentityManager - services that require signing
  identity of the CA registrar

User parameter is removed from Register/Revoke, as it
represents the CA registrar. The registrar information
is read from the CA configuration. If the registrar is
not enrolled yet, SDK will attept to enroll it.

SDK now stores user certs into the user store upon
Enroll/Reenroll.

FabriCAClient will soon be deprecated in favor of
EnrollmentService and IdentityManager that will be
exposed at the top SDK level.

Change-Id: I93961a837544aefd55bdc5ce74d0237c65612f7e
Signed-off-by: Aleksandar Likic <aleksandar.likic@securekey.com>