[FAB-8428] UserStore
Introduced a concept of a UserStore, for persistence of
user information. In this first implementation, only the
user primary key (mspID, name) and enrollment certificate
are stored.

During enrollment, user's private key is added to
cryptosuite store automatically. Application still needs
to add user (that contains the enrollment certificate)
to user store upon enrollment. In the future this won't
be necessary.

Initial implementation of the store is a simple file store
that stores user's enrollment cert in a pem file, using
naming convention <user>@<mspid>-cert.pem. The intention
is to make the user store pluggable in the future, to allow
SDK integrators to plugin their own implementation of
the UserStore interface.

Change-Id: I63452e92a50976b141345a87766541bab2291a8e
Signed-off-by: Aleksandar Likic <aleksandar.likic@securekey.com>