[FAB-8441] Save user to store on Enroll()
Prior to this change, application had to save user
enrollment certificates into a cert store upon
enrollment. This is now done by the enrollment process.
After enrollment, the newly generated private key will
be in the crypto store, and the user enrollment cert will
be in the user store.

Change-Id: Icd433145611b1c2791051a5cbb7f631201953e13
Signed-off-by: Aleksandar Likic <aleksandar.likic@securekey.com>

[FAB-8270] WIP - Split FabriCAClient interface

WIP, don't merge

Change-Id: Idd23b054e85f5010424d1bf88985c723fbe633b4
Signed-off-by: Aleksandar Likic <aleksandar.likic@securekey.com>