[FAB-8717] Request Context
This change adds a Go Context creation method
to the SDK's context package for request scopes. The comm
manager is also added to the request scope. Calls to
orderer and peer methods that Dial are modified to
first create a context using the new context
creation method.

Integration tests were also refactored to use the same
SDK instance for initialiation and usage.

Change-Id: I095016e6da5f8eade13c1f5dcf4d8f9fb287e6ed
Signed-off-by: Troy Ronda <troy@troyronda.com>