[FAB-8899] Dynamic selection service caching peers
In the previous implementation, target peers
were passed into GetEndoresersForChaincode and the
selection service, on first init, cached these
peers and returned them every time thereafter. With
this change the dynamic selection service calls the
discovery service on each call.

The selection service now requires a channel context
in order to retrieve a handle to the discovery service.
A new Initialize function on the selection service is
called by the channel context after it is created in
order to provide it the context.

The peer target filter was also removed from the client
options and into the GetEndorsersForChaincode function.
This filter was really meant to be a per-request filter
as opposed to a per-channel client filter. A filter on
the channel client is not necessary since a custom
discovery provider may be injected with the required
custom filtering logic.

Change-Id: Ib6ad9de332c000c26f3d6d98b425caf1a3614252
Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>