[FAB-9602] entity matcher refactoring
- entity matchers aren't mandatory anymore
- Merged EndpointConfig.PeerConfig and
EndpointConfig.PeerConfigByURL to single function.
- Refactered below config functions,
 * EndpointConfig.PeerConfig - will return nil if
peer config not found instead of throwing error.
* EndpointConfig.PeersConfig - will skip the peer
if corresponding peer config not found, instead of
failing as whole
* EndpointConfig.ChannelPeers - will skip the peer
if corresponding peer config not found, instead of
failing as whole
* EndpointConfig.OrdererConfig - will return nil if
peer config not found instead of throwing any error.
* IdentityConfig CA lookups - will not fail if entity
matchers not found.
- endpoint discovery to skip the peers for which config is
not found, but will fail if in case of error while reading
peer config.





Change-Id: Ieb64514862ab1f3c03878397602d9fee88e71dfa
Signed-off-by: Sudesh Shetty <sudesh.shetty@securekey.com>