[FAB-10622] Add Invocation Chain to Channel Client
Added Invocation Chain meta-data to the Channel Client Request
so that the Selection Service chooses endorsers that satisfy the
endorsement policies of all chaincodes involved in an invocation
chain (i.e. for CC-to-CC invocations). Each chaincode may also
be associated with a set of private data collection names (used
by the Fabric Selection service) to exclude endorsers that do NOT
have read access to the collections.

Change-Id: I561289c941696d66e109621303873f191cffda34
Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>