[FABG-746] Allow transfer of event registrations
A new function was added to the EventClient API that allows
you to transfer event registrations from an existing client
and import them into a new client. The new client will
start listening to blocks from where the old client left off.
This allows you to update the configuration of an SDK without
losing events.

Change-Id: I5f8035b2c29a0a4e10a4f85c8ffa09d4153e04ec
Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>