[FABG-782] Fix default event service policy
If the eventService policy is not defined for a channel
then the one defined under the _default channel is used.

This causes a problem with parameters where 0 is a valid
value, i.e. blockHeightLagThreshold and
reconnectBlockHeightLagThreshold. If 0 is specified
then the SDK will assume that the value was not provided
and then use the default value. For this reason, -2 is now
used for blockHeightLagThreshold to indicate that only
up-to-date peers are to be used and -1 is used for
reconnectBlockHeightLagThreshold to indicate that the
disconnect feature should be disabled.

Change-Id: I03d667548a2649bb719ae756bec59bf533548c57
Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>