[FABG-844] third_party fabric pins as a module
This change converts the third_party fabric code included
in the SDK into a module. This allows substitution of the
imported codebase.

The non-proto Fabric utility code exposed from this module
has also been removed.

A temporary workaround exists in go.mod to enable usage of
the not yet existing third_party fabric module. It will be
removed in the next commit.

Change-Id: Ife3434e8b7656e19f19cf54454a7123a830a91ee
Signed-off-by: Troy Ronda <troy@troyronda.com>