FABG-904 Ability to update channel configuration
This change allows SDK client to update channel configuration
by submitting modified configuration to SDK. SDK will fetch
the latest channel configuration from the orderer, calculate
the config update transaction based on the diff, and submit
the update transaction to the orderer.

WARNING: This change breaks the API for saving channel
configuration. Before the change, type SaveChannelRequest
contained two options for defining the new channel
configuration:

  ChannelConfig     io.Reader // ChannelConfig data source
  ChannelConfigPath string    // Convenience option to use the named file as ChannelConfig reader

The problem with the old API is that the names of the
options don't reflect their meanings. While the names suggest
the new channel configuration, the provided options, in fact,
supply the fabric transcation to be submitted to the orderer
to change channel configuration..

This change adds ChannelConfig option, which is indeed the new
channel configuration, and renames the above two options to
better reflect their meanings. The resulting options are:

  ChannelConfig         *common.Config // Updated channel config.
                                       // If present, takes precedence over ChannelConfigTx
  ChannelConfigTxReader io.Reader      // ChannelConfig Tx data source
  ChannelConfigTxPath   string         // Convenience option to use the named file
                                       // as ChannelConfig Tx reader

Signed-off-by: Aleksandar Likic <aleksandar.likic@securekey.com>
Change-Id: I9a348dd77bfc03f848a10f8dc749b31d987c73a1